Added transposed function for vectors
* Also changed fabs to fabsf in CUDA kernel

+++ src/cuda/kernels/kernels_linear.cu
@@ -129,7 +129,7 @@ void kernel_p_norm_vec(float *lhs, float p, float *rhs) {
   int index = blockDim.x * blockIdx.x + threadIdx.x;
 
   __shared__ float sub_vec[SUB_VECTOR_LEN];
-  sub_vec[threadIdx.x] = powf(fabs(lhs[index]), p);
+  sub_vec[threadIdx.x] = powf(fabsf(lhs[index]), p);
 
   __syncthreads();
